requirements.txt
torch>=2.0.0
transformers>=4.0.0
Pillow>=8.0.0
numpy>=1.19.0
matplotlib>=3.0.0
pandas>=1.2.0
geopandas>=0.10.0
shapely>=1.7.0
scikit-learn>=0.24.0
scikit-image>=0.18.0
rasterio>=1.2.0
requests>=2.25.0
